Add error handling for server responses
This commit adds two handlers. The first handler examines the server
response. If it is not ok (200), this handler throws an `Error` instance
with the details.

The second handler catches all `Errors` and reports them to the user.
Furthermore, this handler only displays the error message for 5 seconds
and then clears it. I find this is not really enough time to process the
error (read and think), but it is only a demonstration.